+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 
 ### Fixed
 - Technical - Remove unused `bcryptjs` dependency.
+- ForestAdmin-Schema - The schema file is now written in the project directory, wherever the startup command has been hinted from.
 
 ## RELEASE 5.2.0 - 2019-11-18
 ### Added